+++ docs/static/1674902298/state.js
@@ -1,0 +1,1 @@
+window.__NUXT__=(function(a,b){return {staticAssetsBase:"https:\u002F\u002Folexandrbig.github.io\u002Ffr-kipa\u002Fstatic\u002F1674902298",layout:"default",error:a,state:a,serverRendered:false,routePath:b,config:{_app:{basePath:"\u002Ffr-kipa\u002F",assetsPath:b,cdnURL:"https:\u002F\u002Folexandrbig.github.io\u002Ffr-kipa\u002F"},content:{dbHash:"8fe72f21"}}}}(null,"\u002F"));