+++ CHANGELOG.md
@@ -8,6 +8,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## 0.12.0 - 2023-01-28
+
+### Added
+* Flows basic view
+* Exception handler operation
+* Designer top actions
+
+
 ## 0.11.1 - 2023-01-20
 
 ### Added